+++ src/assets/scripts/iframefoxit.js
@@ -39,12 +39,14 @@ var foxitViewer = function foxitViewer(zsdivid, divnum, libpath) {
     var foxview = this;
     var PixelToPoint = 72 / 96; //0.75
     var PointToPixel = 96 / 72; //1.3333
-
+        
     //window.onscroll = this.onScroll;
 
     var RxCore = undefined;
 
     window.$ = PDFViewCtrl.jQuery;
+
+    this.zoomfit = false;
 
     this.divnum = divnum;
     this.libpath = libpath;
@@ -1434,6 +1436,7 @@ var foxitViewer = function foxitViewer(zsdivid, divnum, libpath) {
     };    
 
     this.zoomdirect = function (pagenum, factor, scrolldata) {
+        foxview.zoomfit = true;
 
         return new Promise(async (resolve, reject) => {
 
@@ -3565,8 +3568,12 @@ var foxitViewer = function foxitViewer(zsdivid, divnum, libpath) {
 
                 foxview.pagestates[pgindex].rotation = pageRender.page.getRotationAngle();
 
-
-                foxview.setmarkupPositionScale(pgindex);
+                //foxview.setmarkupPositionScale(pgindex);
+                if((foxview.zoomfit) || (!foxview.vBarOn && !foxview.hBarOn) || (foxview.vBarOn && foxview.hBarOn)) {
+                    return;
+                }
+                
+                //foxview.setmarkupPositionScale(pgindex);
 
 
                 var foxitpageEx = {width: pagewidth, height : pageheight, pagescale : pgscale, docindex: foxview.rxindex, pageindex : pgindex, pagepos : undefined};
@@ -3698,9 +3705,9 @@ var foxitViewer = function foxitViewer(zsdivid, divnum, libpath) {
 
                 RxCore.setfoxitPageSize(foxitpage);
 
-                if (foxview.curpage == pgindex) {
+                /*if (foxview.curpage == pgindex) {
                     foxview.setmarkupPosition(foxview.curpage);
-                }
+                }*/
 
                 //console.log('render success called');